% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umich_colors.R
\name{render_umich_colors}
\alias{render_umich_colors}
\title{UMich colors palette rendering}
\usage{
render_umich_colors(name, n)
}
\arguments{
\item{name}{Name of color palette. Choices are:
\code{signature}, \code{secondary}, \code{bills_beer_garden}.}

\item{n}{Number of colors desired. Most palettes have 3-13 colors. Most color
schemes are derived from \href{https://brand.umich.edu/design-resources/colors/}.
If omitted, uses all colors.}
}
\value{
A vector of colors.
}
\description{
UMich colors palette rendering
}
\examples{
render_umich_colors("signature")
render_umich_colors("secondary", 5)
}
\keyword{colors}
