% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umich_colors.R
\docType{data}
\name{umich_palettes}
\alias{umich_palettes}
\title{List of color palettes.}
\format{
An object of class \code{list} of length 3.
}
\usage{
umich_palettes
}
\description{
Use \code{\link{render_umich_colors}} to construct palettes of desired length.
}
\keyword{datasets}
